+++ PolyhydraGames.Ollama/Models/OllamaResponse.cs
@@ -32,7 +32,7 @@ public class OllamaResponse
     public int? PromptEvalCount { get; set; }
 
     [JsonPropertyName("prompt_eval_duration")]
-    public int? PromptEvalDuration { get; set; }
+    public long? PromptEvalDuration { get; set; }
 
     [JsonPropertyName("eval_count")]
     public int? EvalCount { get; set; }
